lava_command: add enter/exit hooks to call host-side tool
This is used to generate measurements by interacting with devices that are
not reachable as a node. Each device.conf file can define a host command or
script used when entering a lava_command_run and leaving a lava_command_run.

host_hook_enter_command = ...(phrase to pass to host)
host_hook_exit_command = ...

Hook calling abi with {enter_hook} as the phrase from the config file above:

ENTRER: arg0 = {enter_hook}  arg1 = logdir/{enter_hook}.dat &
EXIT:   arg0 = {exit_hook}   arg1 = logdir/{enter_hook}.dat &

A test_case_id is created to show the output of the hook as a
'measurement' field (stdout of exit hook)

any file found in the logdir will be attached, starting with
stdout.log. If a {file}.minetype is found, its content is used
as 'mime_type' for the attached file (as with lava_shell_test)

Signed-off-by: Marc Titinger <mtitinger@baylibre.com>