bootloader: add ramdisk_raw option
By default, a uboot header is automatically added to the ramdisk image.

For bootloaders without INITRD_ATAG support, the ramdisk needs to be
passed on the command line and cannot have the uboot header added.

To enable this feature, add a "ramdisk_raw" option that device files can
set so that a uboot header is not added.

Signed-off-by: Kevin Hilman <khilman@baylibre.com>